Started working on application manager, icons by eboye, published version 1.0.1.2

+++ src/rs/pedjaapps/kerneltuner/utility/Tools.java
@@ -65,7 +65,7 @@ public class Tools
 
     }
 
-    public static String kByteToHumanReadableSize(int size)
+    public static String kByteToHumanReadableSize(long size)
     {
         String hrSize = "";
         double m = size / 1024.0;
